fix: preserve decimal points in rubric CSV imports
When importing rubrics via CSV, rating points should be preserved as floating point
numbers rather than being truncated to integers. This maintains consistency with
how points are handled elsewhere in Canvas LMS.

+++ lib/rubric_csv_importer.rb
@@ -67,7 +67,7 @@ class RubricCSVImporter
       {
         description: rating_description,
         long_description: row[rating_indices[:long_description_indices][index]],
-        points: row[rating_indices[:points_indices][index]].to_i
+        points: row[rating_indices[:points_indices][index]].to_f
       }
     end.compact
